backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
mysql-connector-python==8.1.0
PyMySQL==1.1.0

